+++ src/main_page.py
@@ -35,32 +35,36 @@ def main_page(appstate: ChatAppVars):
 
     ###### HEADER ######
     st.write("""<p style="text-align: center; font-size: 60px;">🗣️🤖💬</p>""", unsafe_allow_html=True)
-    if appstate.debug:
-        st.caption(f"session: {appstate.chat.session_start_time}")
-        with st.expander("# Debugging"):
-            st.text_input("mistral API key", appstate.api_key_mistral, key="api_key_mistral", disabled=True)
-            st.text_input("OpenAI API key", appstate.api_key_openai, key="api_key_openai", disabled=True)
-            st.write(f"Description: `{appstate.chat.description}`")
-            st.write(appstate.chat.messages)
+    st.header("", divider="rainbow")
 
-    sidebar(appstate)
+    if st.session_state.get("echobot", False):
+        st.caption("echobot mode")
+
+    if appstate.debug:
+        debugging_placeholder = st.empty()
+
+    # with debugging_placeholder:
+    #     with st.expander("# Debugging"):
+    #         st.write(appstate)
+    #         # st.write(appstate.chat.messages)
+    #         st.write(appstate.chat)
+    #         # st.write("chat_history:", appstate.chat_history)
+
+    with st.sidebar:
+        # vanishing_sidebar = st.container(border=True)
+        # settings_sidebar = st.container(border=False)
+        settings_sidebar = st.empty()
+
+    # settings_sidebar.caption(";)")
+    # settings_sidebar.empty()
+    sidebar(appstate, settings_sidebar)
 
     ###### TOP BUTTONS ######
     st.write(COLUMN_FIX_CSS, unsafe_allow_html=True)
 
     top_buttons = st.columns((2, 1))
     with top_buttons[0]:
-        # st.button("New chat +", on_click=lambda: appstate.new_thread())
         st.empty()
-        # st.button("🌱 New chat", on_click=lambda: appstate.new_thread(), use_container_width=True)
-    # with top_buttons[1]:
-    #     if appstate.chat.messages != []:
-    #         st.button("Delete 🗑️", on_click=delete_this_chat)
-
-    # newbutton, deletebutton= st.columns((1, 1))
-    # newbutton.button("New chat +", on_click=lambda: appstate.new_thread(), use_container_width=True, key="button_new")
-
-    # st.write("---")
 
     ####### CONVERSATION #######
     for message in appstate.chat.messages:
@@ -83,22 +87,42 @@ def main_page(appstate: ChatAppVars):
         from streamlit_mic_recorder import speech_to_text
         prompt = speech_to_text( language='en', use_container_width=True, just_once=True, key='STT')
 
+    # st.toggle("🔊", key="toggle_sound")
     if prompt:
+        settings_sidebar.empty()
+        # disable 'echobot' checkbox
+        # st.session_state.echobot.disabled = True
+
+        # settings_sidebar = st.container()
+        # st.session_state.incomplete_stream = ""
+        with st.sidebar:
+            # with vanishing_sidebar:
+            st.button("Interrupt", on_click=interrupt, key="button_interrupt")
+
         my_next_prompt.chat_message("user").markdown(prompt)
-        run_prompt(prompt, bots_reply)
-        save_chat_history()
+        reply = run_prompt(prompt, bots_reply)
 
-    if appstate.chat.not_yet_saved:
-        # A new chat thread has just been created, so we must update our list of past conversations
-        appstate.load_chat_history()
+        new_chat = save_chat_history() # dummy variable for readability
+        if new_chat:
+            # A new chat thread has just been created, so we must update our list of past conversations
+            appstate.load_chat_history()
 
-    with st.sidebar:
-        st.write("---")
-        # center_text("h3", "Past conversations")
-        # with st.expander("Past conversations", expanded=not appstate.debug):
-        with st.expander("Past conversations", expanded=False):
-            for description, runlog in appstate.chat_history:
-                st.button(f"{description}", on_click=load_convo, args=(runlog,), use_container_width=True, key=runlog.split('.')[0])
+        if 'read_to_me' in st.session_state and st.session_state.read_to_me == True:
+            st.session_state.tts = reply
+
+        st.rerun()
+        # sidebar(appstate, settings_sidebar) # this will cause a KeyError because of dupicate key widget keys
+
+    if 'tts' not in st.session_state:
+        st.session_state.tts = None
+
+    if st.session_state.tts is not None:
+        TTS(st.session_state.tts)
+        st.session_state.tts = None
+        # TTS(reply) # may throw an exception
+
+    #### AFTER-PROMPT PROCESSING ####
+    # put things here to update the UI _AFTER_ the prompt has been run
 
     if len(appstate.chat.messages) > 0:
         with top_buttons[0]:
@@ -106,10 +130,23 @@ def main_page(appstate: ChatAppVars):
         with top_buttons[1]:
             st.button("🗑️ Delete", on_click=delete_this_chat, key="button_delete")
 
+
     with st.sidebar:
+        with st.expander("Past conversations", expanded=False):
+            for description, runlog in appstate.chat_history:
+                st.button(f"{description}", on_click=load_convo, args=(runlog,), use_container_width=True, key=runlog.split('.')[0])
+
         st.write("---")
         st.button(f"⚙️ Configure", on_click=settings, use_container_width=True)
         st.caption(f"Logged in as: `{st.session_state.appstate.username}`")
+
+
+    if appstate.debug:
+        with debugging_placeholder:
+            with st.expander("# Debugging"):
+                st.write(appstate)
+                st.write(appstate.chat)
+                # st.write(appstate.chat.messages)
 
 
     ### outside of the if prompt block
@@ -120,15 +157,20 @@ def settings():
     st.session_state.route = PageRoute.SETTINGS
 
 
-
-def sidebar(appstate):
+def sidebar(appstate, settings_sidebar):
     #Note: we do this at the end so that a new chat history will be displayed after the users first message
-    with st.sidebar:
+    
+    # with st.sidebar:
+    # with settings_sidebar:
+    with settings_sidebar.container(border=True):
+        if appstate.debug:
+            # st.checkbox("Echobot", key="echobot", value=appstate.debug)
+            st.toggle("Echobot", key="echobot", value=appstate.debug)
         st.radio("Input method:", ["Text ⌨️", "Voice 🗣️"], index=0, key="input_method")
         if appstate.api_key_openai in [None, ""]:
             st.info("Enter OpenAI key in settings to enable text-to-speech")
         else:
-            st.checkbox("Read aloud 👂", key="read_to_me", value=False, on_change=None) #on_change must be None or else a re-run of last prompt happens
+            st.toggle("Read aloud 👂", key="read_to_me", value=False, on_change=None) #on_change must be None or else a re-run of last prompt happens
 
         if 'read_to_me' in st.session_state and st.session_state.read_to_me == True:
 
@@ -136,13 +178,14 @@ def sidebar(appstate):
             cols[0].radio("Voice model", TTS_VOICE_CHOICES, index=1, key="openai_voice")
             cols[1].radio("Talking speed", [1.0, 1.2, 1.5], index=1, key="tts_rate")
         st.write("---")
+        st.toggle("Safe mode", key="mistrel_safemode", value=False, help="Safe mode is not yet implemented by mistral ai", disabled=True)
         st.radio("Model",
-                 appstate.mistral_models,
-                 index=2 if appstate.debug else 0,
-                 key="mistrel_model")
-        st.checkbox("Safe mode", key="mistrel_safemode", value=False, help="Safe mode is not yet implemented by mistral ai", disabled=True)
+                appstate.mistral_models,
+                index=2 if appstate.debug else 0,
+                key="mistrel_model")
 
-
+        # vanishing_sidebar = st.empty()
+    # return vanishing_sidebar
 
 
 def run_prompt(prompt, bots_reply):
@@ -150,37 +193,38 @@ def run_prompt(prompt, bots_reply):
 
     # With streaming
     with bots_reply.chat_message("assistant"):
-        incomplete_stream = ""
+        st.session_state.incomplete_stream = ""
         place_holder = st.empty()
 
         try:
             with st.spinner("🧠 Thinking..."):
-                client = st.session_state.appstate.get_client()
+                try:
+                    client = st.session_state.appstate.get_client()
+                except Exception as e:
+                    # TODO missing API keys will throw an exception here
+                    # WE NEED TO AVOID THIS by limiting the options in the sidebar - don't let the user pick a model that won't work!
+                    st.error(e)
+                    st.stop()
+
                 # for chunk in st.session_state.appstate.get_client():
                 for chunk in client:
                     try:
-                        incomplete_stream += chunk.choices[0].delta.content
+                        st.session_state.incomplete_stream += chunk.choices[0].delta.content
                     except TypeError:
                         #TODO - not sure why this error happens...
                         # st.session_state.incomplete_stream += chunk.choices[0].message.content
                         print("TypeError in run_prompt()")
                         pass
-                    place_holder.markdown(incomplete_stream)
+                    place_holder.markdown(st.session_state.incomplete_stream)
         except MistralAPIException as e:
             st.error(e)
             st.stop()
 
-        reply = incomplete_stream
+        reply = st.session_state.incomplete_stream
 
     st.session_state.appstate.chat.messages.append(ChatMessage(role="assistant", content=reply))
 
-    # if st.session_state.output_method == "Voice":
-    if 'read_to_me' in st.session_state and st.session_state.read_to_me == True:
-        TTS(reply) # may throw an exception
-
-
-
-
+    return reply
 
 
 # def autoplay_audio(file_path: str):
@@ -224,8 +268,6 @@ def TTS(text, language='en', slow=False):
             except gTTSError as e:
                 st.error(e)
                 st.error(f"Could not create audio: ")
-                # st.stop()
-
 
         else:
             # if "openai_client" not in st.session_state:
@@ -259,3 +301,16 @@ def TTS(text, language='en', slow=False):
             # Convert audio data to Base64
             audio_base64 = base64.b64encode(audio_data).decode('utf-8')
             autoplay_audio(audio_base64)
+
+
+def interrupt():
+    # st.session_state.interrupt = True
+    st.session_state.appstate.chat.messages.append(ChatMessage(role="assistant", content=st.session_state.incomplete_stream))
+    st.session_state.appstate.chat.messages.append(ChatMessage(role="user", content="<INTERRUPTS>"))
+    # save_chat_history()
+    # st.session_state.appstate.load_chat_history()
+
+    if save_chat_history():
+        st.session_state.appstate.load_chat_history()
+
+    # st.rerun() # not allowed in on_click handlers (callbacks)